MMS: Fix Application crash: WorkingMessage.asyncUpdateDraftMmsMessage
After we can forward unsupported mms attachment, the unsupported type may
have an unexpected uri, such as missing authority part. And this kind of
uri will casue NullPointerException in PduPersister. So we need to make
sure the uri have an authority before use it.

Check the unsupported type's uri to avoid NullPointerException.

CRs-Fixed: 583527

Change-Id: I5f83b1d15b6ada05a7ec28e7f20eb1cabb33672b